requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
python-multipart>=0.0.6

# Web scraping
beautifulsoup4>=4.12.0
playwright>=1.41.0

# Document processing
pypdf>=4.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
pandas>=2.2.0

# Optional: Vector store for semantic search
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Development
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
